server/requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
